Added swipe and weather for next days
