+++ WeatherInfo/Controllers/WeatherListVC.swift
@@ -8,61 +8,53 @@
 import UIKit
 
 class WeatherListVC: UIViewController {
-    
-    @IBOutlet weak var weatherTableView: UITableView!
+
+    @IBOutlet weak var weatherCollectionView: UICollectionView!
     private var arrayLocations: [WeatherViewModel] = []
-    private let cellID = "WeatherTableViewCell"
+    private let headerId = "WeatherCVHeaderView"
+    private let cellId = "WeatherCVCell"
     private var weatherManager = WeatherManager()
     
-    // MARK:- LifeCycle
     override func viewDidLoad() {
         super.viewDidLoad()
-        weatherManager.delegate = self
         setupUI()
-    }
-    
-    private func setupUI() {
-        let nib = UINib(nibName: cellID, bundle: nil)
-        weatherTableView.register(nib, forCellReuseIdentifier: cellID)
-        loadTableData()
         loadLocationData()
     }
     
-    // MARK:- Table View Observer
-    private func loadTableData() {
-//        DispatchQueue.main.asyncAfter(deadline: .now()+1, execute: {
-//            self.arrayLocations.asObservable().bind(to: self.weatherTableView.rx.items(cellIdentifier: self.cellID, cellType: WeatherTableViewCell.self)) { index,item,cell in
-//                cell.updateCell(with: item as AnyObject)
-//            }.disposed(by: self.disposeBag)
-//        })
+    func setupUI() {
+        let nib = UINib(nibName: cellId, bundle: nil)
+        weatherCollectionView.register(nib, forCellWithReuseIdentifier: cellId)
+        let headerNib = UINib(nibName: headerId, bundle: nil)
+        weatherCollectionView.register(headerNib, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: headerId)
     }
-    
+
     // MARK:- Fetch Given cities weather condition
     private func loadLocationData() {
-        weatherManager.fetchBulkCitiesWeather()
+        weatherManager.delegate = self
+        weatherManager.fetchBulkWeather()
     }
-    
+
     // MARK:- Alerts
-    
     private func showAlert(_ msg: String) {
         DispatchQueue.main.asyncAfter(deadline: .now()+1) {
             Alert.showAlert(message: msg)
         }
     }
 }
-
 //MARK: - WeatherManagerDelegate
 
 extension WeatherListVC: WeatherManagerDelegate {
     
     func didUpdateWeather(_ weatherManager: WeatherManager, weather: WeatherViewModel) {
-//        arrayLocations = weather
+        arrayLocations.append(weather)
+        weatherCollectionView.reloadData()
     }
     
     func didUpdateBulkWeather(_ weatherManager: WeatherManager, weather: [WeatherViewModel]) {
         arrayLocations = weather
+        print(arrayLocations)
         DispatchQueue.main.async {
-            self.weatherTableView.reloadData()
+            self.weatherCollectionView.reloadData()
         }
     }
     
@@ -71,34 +63,27 @@ extension WeatherListVC: WeatherManagerDelegate {
     }
 }
 
-//MARK:- Tableview delegate and datasource methods
-
-extension WeatherListVC: UITableViewDataSource, UITableViewDelegate {
-
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return arrayLocations.count
+extension WeatherListVC: UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
+    
+    ///Number of Sections in Collection View
+    func numberOfSections(in collectionView: UICollectionView) -> Int {
+        return 1
+    }
+    ///Number of Rows in Sections in Collection View
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return 3
     }
 
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: cellID, for: indexPath) as! WeatherTableViewCell
-        cell.updateCell(with: arrayLocations[indexPath.row])
+    /// Collection Cell Size
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        let width = weatherCollectionView.frame.width
+        return CGSize(width: width, height: width * 1.8)
+    }
+
+    /// Cell For Row at
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "WeatherCVCell", for: indexPath) as! WeatherCVCell
+        cell.updateUI(model: arrayLocations)
         return cell
-    }
-    
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let passvc = StoryBoardMain.detailVC.instantiate()
-//        passvc.weather = arrayLocations.value[indexPath.row]
-        self.present(passvc, animated: true, completion: nil)
-    }
-    
-    func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
-        if indexPath.row == 0 { return nil }
-        let deleteAction = UIContextualAction(style: .destructive, title: Constants.strDelete) { (action, view, handler) in
-            self.arrayLocations.remove(at: indexPath.row)
-        }
-        deleteAction.backgroundColor = .red
-        let configuration = UISwipeActionsConfiguration(actions: [deleteAction])
-        configuration.performsFirstActionWithFullSwipe = false
-        return configuration
     }
 }
